Added ts for components view meeting
